# apps/api/src/routes/products.ts
import { Router } from "express";
import { and, eq, ilike, isNull, sql } from "drizzle-orm";
import { product, withTenant, type Tx } from "@rmixerp/db";
import { CreateProductBody, UpdateProductBody, VoidProductBody, type Product } from "@rmixerp/contract";
import { db } from "../db";
import { requireAuth } from "../middleware/requireAuth";
import { requirePermission } from "../middleware/requirePermission";
import { parsePagination, paginatedBody } from "../lib/pagination";
import { sendCsv } from "../lib/csv";
import { notFound, validationError } from "../lib/errors";
import { paramId, queryString } from "../lib/params";
import { writeAudit } from "../audit";

export const productsRouter = Router();
const MODULE = "products";

type ProductRow = typeof product.$inferSelect;

function toApi(row: ProductRow): Product {
  return {
    id: row.id,
    companyId: row.companyId,
    branchId: row.branchId,
    name: row.name,
    code: row.code,
    characteristicStrengthMpa: row.characteristicStrengthMpa,
    unit: row.unit,
    description: row.description,
    createdAt: row.createdAt.toISOString(),
    updatedAt: row.updatedAt.toISOString(),
    voidedAt: row.voidedAt?.toISOString() ?? null,
  };
}

async function findActive(tx: Tx, id: string): Promise<ProductRow | undefined> {
  const [row] = await tx
    .select()
    .from(product)
    .where(and(eq(product.id, id), isNull(product.voidedAt)));
  return row;
}

productsRouter.get("/products", requireAuth, requirePermission(MODULE, "view"), async (req, res) => {
  const pagination = parsePagination(req);
  const q = queryString(req, "q");

  const { items, total } = await withTenant(db, req.auth!.companyId, async (tx) => {
    const where = and(isNull(product.voidedAt), q ? ilike(product.name, `%${q}%`) : undefined);
    const [rows, countRows] = await Promise.all([
      tx.select().from(product).where(where).limit(pagination.limit).offset(pagination.offset),
      tx.select({ count: sql<number>`count(*)::int` }).from(product).where(where),
    ]);
    return { items: rows, total: countRows[0]?.count ?? 0 };
  });

  if (queryString(req, "format") === "csv") {
    sendCsv(res, "products.csv", items.map(toApi), [
      { key: "id", header: "ID" },
      { key: "name", header: "Name" },
      { key: "code", header: "Code" },
      { key: "characteristicStrengthMpa", header: "Strength (MPa)" },
      { key: "unit", header: "Unit" },
    ]);
    return;
  }
  res.status(200).json(paginatedBody(items.map(toApi), total, pagination));
});

productsRouter.get("/products/:id", requireAuth, requirePermission(MODULE, "view"), async (req, res) => {
  const row = await withTenant(db, req.auth!.companyId, (tx) => findActive(tx, paramId(req)));
  if (!row) {
    res.status(404).json(notFound("Product"));
    return;
  }
  res.status(200).json(toApi(row));
});

productsRouter.post("/products", requireAuth, requirePermission(MODULE, "create"), async (req, res) => {
  const parsed = CreateProductBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json(validationError(parsed.error.flatten()));
    return;
  }
  const input = parsed.data;

  const created = await withTenant(db, req.auth!.companyId, async (tx) => {
    const [row] = await tx
      .insert(product)
      .values({
        companyId: req.auth!.companyId,
        branchId: input.branchId ?? null,
        name: input.name,
        code: input.code,
        characteristicStrengthMpa: input.characteristicStrengthMpa ?? null,
        unit: input.unit ?? "m3",
        description: input.description ?? null,
        createdBy: req.auth!.userId,
      })
      .returning();
    if (!row) throw new Error("insert returned no row");
    await writeAudit(tx, {
      companyId: req.auth!.companyId,
      actorUserId: req.auth!.userId,
      entityType: "product",
      entityId: row.id,
      action: "create",
      after: row,
    });
    return row;
  });

  res.status(201).json(toApi(created));
});

productsRouter.put("/products/:id", requireAuth, requirePermission(MODULE, "edit"), async (req, res) => {
  const parsed = UpdateProductBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json(validationError(parsed.error.flatten()));
    return;
  }
  const input = parsed.data;
  const id = paramId(req);

  const updated = await withTenant(db, req.auth!.companyId, async (tx) => {
    const before = await findActive(tx, id);
    if (!before) return null;

    const [row] = await tx
      .update(product)
      .set({
        ...(input.branchId !== undefined && { branchId: input.branchId }),
        ...(input.name !== undefined && { name: input.name }),
        ...(input.code !== undefined && { code: input.code }),
        ...(input.characteristicStrengthMpa !== undefined && {
          characteristicStrengthMpa: input.characteristicStrengthMpa,
        }),
        ...(input.unit !== undefined && { unit: input.unit }),
        ...(input.description !== undefined && { description: input.description }),
        updatedAt: new Date(),
      })
      .where(eq(product.id, id))
      .returning();
    if (!row) throw new Error("update returned no row");
    await writeAudit(tx, {
      companyId: req.auth!.companyId,
      actorUserId: req.auth!.userId,
      entityType: "product",
      entityId: row.id,
      action: "update",
      before,
      after: row,
    });
    return row;
  });

  if (!updated) {
    res.status(404).json(notFound("Product"));
    return;
  }
  res.status(200).json(toApi(updated));
});

productsRouter.delete("/products/:id", requireAuth, requirePermission(MODULE, "void"), async (req, res) => {
  const parsed = VoidProductBody.safeParse(req.body ?? {});
  const reason = parsed.success ? (parsed.data.reason ?? null) : null;
  const id = paramId(req);

  const voided = await withTenant(db, req.auth!.companyId, async (tx) => {
    const before = await findActive(tx, id);
    if (!before) return null;

    const [row] = await tx.update(product).set({ voidedAt: new Date() }).where(eq(product.id, id)).returning();
    if (!row) throw new Error("void returned no row");
    await writeAudit(tx, {
      companyId: req.auth!.companyId,
      actorUserId: req.auth!.userId,
      entityType: "product",
      entityId: row.id,
      action: "void",
      before,
      after: row,
      reason,
    });
    return row;
  });

  if (!voided) {
    res.status(404).json(notFound("Product"));
    return;
  }
  res.status(204).send();
});