# packages/contract/openapi.yaml
openapi: 3.1.0
info:
  title: RMC ERP API
  version: 0.2.0
  description: >
    Contract-first source of truth. The TS client, React Query hooks, and
    Zod request/response schemas consumed by apps/web and apps/mobile are
    generated from this file — see CLAUDE.md. Do not hand-write a fetch
    call that bypasses the generated hooks.

    Money fields are JOD decimal strings (e.g. "1234.567"), never numbers
    — formatting/parsing to the internal integer-fils representation
    happens at this HTTP boundary (packages/core's jodStringToFils /
    filsToJodString), matching CLAUDE.md's money conventions.
servers:
  - url: /api
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/QParam"
        - $ref: "#/components/parameters/FormatParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductListResponse" }
            text/csv:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProductRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /products/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by id.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProductRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/ValidationError" }
    delete:
      operationId: voidProduct
      tags: [products]
      summary: Soft-delete (void) a product.
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VoidRequest" }
      responses:
        "204": { description: Voided }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /raw-materials:
    get:
      operationId: listRawMaterials
      tags: [rawMaterials]
      summary: List raw materials.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/QParam"
        - $ref: "#/components/parameters/FormatParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RawMaterialListResponse" }
            text/csv:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createRawMaterial
      tags: [rawMaterials]
      summary: Create a raw material.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRawMaterialRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RawMaterial" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /raw-materials/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      operationId: getRawMaterial
      tags: [rawMaterials]
      summary: Get a raw material by id.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RawMaterial" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: updateRawMaterial
      tags: [rawMaterials]
      summary: Update a raw material.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateRawMaterialRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RawMaterial" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/ValidationError" }
    delete:
      operationId: voidRawMaterial
      tags: [rawMaterials]
      summary: Soft-delete (void) a raw material.
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VoidRequest" }
      responses:
        "204": { description: Voided }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1 }
    PageSizeParam:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100 }
    QParam:
      name: q
      in: query
      description: Free-text search.
      schema: { type: string }
    FormatParam:
      name: format
      in: query
      schema: { type: string, enum: [json, csv] }
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Missing permission
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, code]
          properties:
            message: { type: string }
            code: { type: string }
            details: { type: object }
    VoidRequest:
      type: object
      properties:
        reason: { type: string }
...
    Product:
      type: object
      required:
        [id, companyId, branchId, name, code, characteristicStrengthMpa, unit, description,
         createdAt, updatedAt, voidedAt]
      properties:
        id: { type: string, format: uuid }
        companyId: { type: string, format: uuid }
        branchId: { type: string, format: uuid, nullable: true }
        name: { type: string }
        code: { type: string }
        characteristicStrengthMpa: { type: integer, nullable: true }
        unit: { type: string }
        description: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        voidedAt: { type: string, format: date-time, nullable: true }
    CreateProductRequest:
      type: object
      required: [name, code]
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1, maxLength: 32 }
        branchId: { type: string, format: uuid, nullable: true }
        characteristicStrengthMpa: { type: integer, nullable: true }
        unit: { type: string }
        description: { type: string, nullable: true }
    UpdateProductRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1, maxLength: 32 }
        branchId: { type: string, format: uuid, nullable: true }
        characteristicStrengthMpa: { type: integer, nullable: true }
        unit: { type: string }
        description: { type: string, nullable: true }
    ProductListResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Product" } }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...
    RawMaterial:
      type: object
      required: [id, companyId, branchId, name, code, unit, category, createdAt, updatedAt, voidedAt]
      properties:
        id: { type: string, format: uuid }
        companyId: { type: string, format: uuid }
        branchId: { type: string, format: uuid, nullable: true }
        name: { type: string }
        code: { type: string }
        unit: { type: string }
        category: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        voidedAt: { type: string, format: date-time, nullable: true }
    CreateRawMaterialRequest:
      type: object
      required: [name, code, unit]
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1, maxLength: 32 }
        unit: { type: string, minLength: 1 }
        branchId: { type: string, format: uuid, nullable: true }
        category: { type: string, nullable: true }
    UpdateRawMaterialRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1, maxLength: 32 }
        unit: { type: string, minLength: 1 }
        branchId: { type: string, format: uuid, nullable: true }
        category: { type: string, nullable: true }
    RawMaterialListResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/RawMaterial" } }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...